+++ fractal_server/tasks/utils.py
@@ -1,4 +1,3 @@
-import re
 from pathlib import Path
 
 from fractal_server.config import get_settings
@@ -42,38 +41,8 @@ def get_collection_log_v1(path: Path) -> str:
     return log
 
 
-def get_collection_log_v2(path: Path) -> str:
-    log_path = get_log_path(path)
-    log = log_path.open().read()
-    return log
-
-
 def get_collection_freeze_v1(venv_path: Path) -> str:
     package_path = get_absolute_venv_path_v1(venv_path)
     freeze_path = get_freeze_path(package_path)
     freeze = freeze_path.open().read()
     return freeze
-
-
-def get_collection_freeze_v2(path: Path) -> str:
-    freeze_path = get_freeze_path(path)
-    freeze = freeze_path.open().read()
-    return freeze
-
-
-def _normalize_package_name(name: str) -> str:
-    """
-    Implement PyPa specifications for package-name normalization
-
-    The name should be lowercased with all runs of the characters `.`, `-`,
-    or `_` replaced with a single `-` character. This can be implemented in
-    Python with the re module.
-    (https://packaging.python.org/en/latest/specifications/name-normalization)
-
-    Args:
-        name: The non-normalized package name.
-
-    Returns:
-        The normalized package name.
-    """
-    return re.sub(r"[-_.]+", "-", name).lower()